+++ lua/plugins/marks.lua
@@ -1,0 +1,9 @@
+-- lua/plugins/marks.lua
+
+return {
+    'chentoast/marks.nvim',
+    event = 'BufReadPost',
+    config = function()
+        require('config.marks')
+    end,
+}